Fixing Erik's Mistakes
Fixed the API in the ap.py file.  It actually wasn't an API now it is.  So you can get the lat / lng for all the data now at /api/location.  Should now be able to do a GET request for that too and pull into into Javascript or something else.  If this isn't working for someone please let me know.  I also updated the /data so that the template is called data.html to avoid confusion with our index.html Daniel is making.