Reset of the project 10/23/2018
Good news, the map works!  I removed the duplicate points and got it work.  It was a lot of work but it works. Bad news, although I think Naazneen's files are in the right places they aren't working. Its an integration problem not a Naazneen problem.  If Naazneen could work on this that would be awesome.  We also need Daniel's main index page.  The map is currently standing in for that but the map should be on the index page rather than being the index page.   Lots of work to do between now and Saturday, we also need to start working on the presentation.  All, if you do make changes between now and class on Wednesday DO NOT PUSH TO THE MASTER.  Thanks!  Just bring your files to class and we will work on integration.